gpapi/requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.15.1
charset-normalizer==3.4.1
cryptography==44.0.2
google==3.0.0
idna==3.10
protobuf==4.24.4
pycparser==2.21
python-dotenv==0.21.1
requests==2.28.2
soupsieve==2.4.1
typing-extensions==4.7.1
urllib3==1.26.20
